log-analysis-tool/log-analysis-tool-smolagent/requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0.0
streamlit>=1.27.0
python-dotenv>=1.0.0
numpy>=1.24.0
regex>=2023.0.0
tqdm>=4.65.0
colorama>=0.4.6 
setuptools
smolagents>=1.14.0
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
litellm>=1.9.0